Parameter old-api takes precedence over Maven settings
The swagger brake maven plugin sets default values for the groupId, artifactId, and currentVersion parameters.
This causes an error when using the oldApi parameter from the Maven plugin.
Rather than removing the default values from the Maven plugin and introducing breaking changes, we are updating the logic in the core library.
This should not introduce breaking changes because current users of the library use either the old-api param or the Maven settings.